requirements.txt
django==2.2.3
djangorestframework==3.9.4
djangorestframework_simplejwt==4.3.0
django-rest-auth==0.9.5
python-dotenv==0.10.3
mysqlclient==1.4.2.post1
gunicorn==19.9.0
drf-yasg==1.16.0
factory_boy==2.12.0
django-debug-toolbar==2.0
django-modeladmin-reorder==0.3.1
